restructure project, add more documentation
