+++ stairs/core/pipeline/pipeline_objects/context.py
