+++ stairs/core/pipeline/pipeline_objects/transformation.py
