--- stairs/core/producer/adapter/iter_worker_adapter/utils/add_jobs.lua
@@ -1,11 +1,0 @@
-
-
-local step_key = KEYS[1]
-
-
-for i, value in ipairs(ARGV) do
-    redis.call("lpush", step_key, value)
-end
-
-
-
